Change multiply to xorps or blendps in some cases
Usage of mulps to negate can be replaced with xorps, which can execute
on more ports (SKL and newer), and has a lower latency.

Usage of mulss by zero to zero out just the lowest lane can be replaced
by blendps, which can execute on more ports (Haswell and newer) and has
a lower latency.